+++ models/marts/assessments/rpt_student_assessment.sql
@@ -11,7 +11,9 @@ WITH assessments AS (
             )
         ) AS assessment_student_score
     FROM {{ ref('fct_student_assessment') }} fct_student_assessment
-    WHERE fct_student_assessment.objective_assessment_key = ""
+    LEFT JOIN {{ ref('dim_assessment') }} dim_assessment
+        ON fct_student_assessment.assessment_key = dim_assessment.assessment_key
+    WHERE dim_assessment.objective_assessment_identification_code IS NULL
     GROUP BY 1, 2
 
 ),
@@ -23,32 +25,24 @@ objective_assessments AS (
         fct_student_assessment.student_assessment_identifier,
         ARRAY_AGG(
             STRUCT(
-                dim_objective_assessment.identification_code              AS identification_code,
-                dim_objective_assessment.description                      AS description,
-                fct_student_assessment.reporting_method                   AS reporting_method,
-                fct_student_assessment.student_result                     AS student_result
+                dim_assessment.objective_assessment_identification_code              AS identification_code,
+                dim_assessment.objective_assessment_description                      AS description,
+                fct_student_assessment.reporting_method                                        AS reporting_method,
+                fct_student_assessment.student_result                                          AS student_result
             )
         ) AS objective_assessment_student_score
     FROM {{ ref('fct_student_assessment') }} fct_student_assessment
-    LEFT JOIN {{ ref('dim_objective_assessment') }} dim_objective_assessment
-        ON fct_student_assessment.assessment_key = dim_objective_assessment.assessment_key
-        AND fct_student_assessment.objective_assessment_key = dim_objective_assessment.objective_assessment_key
-        AND fct_student_assessment.school_year = dim_objective_assessment.school_year
-    WHERE fct_student_assessment.objective_assessment_key != ""
+    LEFT JOIN {{ ref('dim_assessment') }} dim_assessment
+        ON fct_student_assessment.assessment_key = dim_assessment.assessment_key
+    WHERE dim_assessment.objective_assessment_identification_code IS NOT NULL
     GROUP BY 1, 2
 
 )
 
 SELECT
     fct_student_assessment.school_year                          AS school_year,
-    dim_assessment.assessment_identifier                        AS assessment_identifier,
-    dim_assessment.title                                        AS title,
-    dim_assessment.namespace                                    AS namespace,
-    dim_student.student_unique_id                               AS student_unique_id,
-    fct_student_assessment.student_assessment_identifier        AS student_assessment_identifier,
-    objective_assessments.objective_assessment_student_score    AS objective_assessment_student_score,
-    assessments.assessment_student_score                        AS assessment_student_score,
     dim_school.school_name                                      AS school_name,
+    dim_student.student_unique_id                               AS student_unique_id,
     dim_student.student_last_surname                            AS student_last_surname,
     dim_student.student_first_name                              AS student_first_name,
     dim_student.student_display_name                            AS student_display_name,
@@ -60,7 +54,13 @@ SELECT
     dim_student.is_english_language_learner                     AS is_english_language_learner,
     dim_student.in_special_education_program                    AS in_special_education_program,
     dim_student.is_hispanic                                     AS is_hispanic,
-    dim_student.race_and_ethnicity_roll_up                      AS race_and_ethnicity_roll_up
+    dim_student.race_and_ethnicity_roll_up                      AS race_and_ethnicity_roll_up,
+    dim_assessment.assessment_identifier                        AS assessment_identifier,
+    dim_assessment.title                                        AS title,
+    dim_assessment.namespace                                    AS namespace,
+    fct_student_assessment.student_assessment_identifier        AS student_assessment_identifier,
+    objective_assessments.objective_assessment_student_score    AS objective_assessment_student_score,
+    assessments.assessment_student_score                        AS assessment_student_score
 FROM {{ ref('fct_student_assessment') }} fct_student_assessment
 LEFT JOIN {{ ref('dim_assessment') }} dim_assessment
     ON fct_student_assessment.assessment_key = dim_assessment.assessment_key
@@ -72,6 +72,6 @@ LEFT JOIN {{ ref('dim_student') }} dim_student
     ON fct_student_assessment.student_key = dim_student.student_key
 LEFT JOIN {{ ref('dim_school') }} dim_school
     ON fct_student_assessment.school_key = dim_school.school_key
-WHERE
-    fct_student_assessment.objective_assessment_key = ""
+WHERE 
+    dim_assessment.objective_assessment_identification_code IS NULL
     AND dim_student.student_unique_id IS NOT NULL